Bug 2180664: optimize metrics tab with multiple components
